fix: ajuste estrutura do server
