+++ server/src/middlewares/error.middleware.ts
@@ -2,9 +2,11 @@
  * Classe base para todos os erros da aplicação.
  */
 export class AppError extends Error {
-  constructor(message: string) {
+  public statusCode: number;
+
+  constructor(message: string, statusCode = 400) {
     super(message);
-    this.name = 'AppError';
+    this.statusCode = statusCode;
   }
 }
