+++ server/src/modules/stores/store.routes.ts
