+++ server/src/modules/stores/store.service.ts
@@ -1,6 +1,6 @@
-import { StoresRepository } from "../repositories/stores.repository.js";
+import { StoresRepository } from "./repositories/stores.repository.js";
 import { TeamsRepository } from "../../teams/repositories/teams.repository.js";
-import { AppError } from "../../../utils/appError.utils.js";
+import { AppError } from "../../middlewares/error.middleware.js";
 
 export class StoresService {
   private repo = new StoresRepository();
