+++ server/src/modules/stores/stores.dtos.ts
