--- server/src/utils/appError.utils.ts
@@ -1,8 +1,0 @@
-export class AppError extends Error {
-  public statusCode: number;
-
-  constructor(message: string, statusCode = 400) {
-    super(message);
-    this.statusCode = statusCode;
-  }
-}
